requirements.txt
requests==2.31.0
pandas==2.0.3
folium==0.14.0
matplotlib==3.7.2
plotly==5.15.0
flask==2.3.2
werkzeug==2.3.6
python-dotenv==1.0.0
geopy==2.3.0
numpy==1.24.3
stravalib==1.3.0
seaborn==0.12.2
